requirements.txt
Django==3.2.5
django-cleanup==5.2.0
django-hashid-field==3.2.1
hurry.filesize==0.9
